+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add convenience `Color::linearRgb()` getter
 - Add interpolation in `Linear RGB` for `Color`
 - Add gamma correction section and example to README
+- Add Björn Ottoson's Oklab perceptual color space
 - Clarify `Color::Space::Rgb` usage and default `sRGB` working space
 
 ### Changed